Huge rework & refactoring.
- Removed unused files.

- Implementations now in .cpp files.

- Some documentation (should i integrate doxygen?)

- Fixes.

+++ engine/src/core/components/pixel_camera.cpp
@@ -19,6 +19,6 @@ float a2d::PixelCamera::GetScaleFactor() {
 }
 
 void a2d::PixelCamera::Update() {
-    SetHeight(a2d::Renderer::GetHeight());
+    SetHeight((float)a2d::Renderer::GetHeight());
     Component::Update();
 }
